requirements.txt
fastapi 
uvicorn 
requests
faiss-cpu 
numpy
sentence-transformers
groq
lxml
python-dotenv